+++ config/modules.config.php
@@ -6,6 +6,7 @@
  * This should be an array of module namespaces used in the application.
  */
 return [
+    'Zend\Navigation',
     'Zend\Session',
     'Zend\Db',
     'Zend\Form',