+++ module/Application/src/Module.php
@@ -2,9 +2,14 @@
 
 namespace Application;
 
+use Zend\Authentication\AuthenticationServiceInterface;
+use Zend\Db\Adapter\AdapterInterface;
 use Zend\Mvc\MvcEvent;
+use Zend\Navigation;
+use Zend\Permissions\Acl\AclInterface;
 use Zend\Session\Exception\RuntimeException as SessionValidationException;
 use Zend\Session\SessionManager;
+use Zend\View\Helper\Navigation as NavigationHelper;
 
 class Module
 {
@@ -15,10 +20,11 @@ class Module
         $application = $e->getApplication();
         $eventManager = $application->getEventManager();
         $serviceManager = $application->getServiceManager();
+        $db = $serviceManager->get(AdapterInterface::class);
 
         $errorListener = new ErrorListener(
             $serviceManager->get('config'),
-            $serviceManager->get('Zend\Db\Adapter\AdapterInterface')
+            $db
         );
         $errorListener->attach($eventManager);
 
@@ -31,6 +37,33 @@ class Module
             $sessionManager->destroy();
             $sessionManager->getValidatorChain()->clearListeners('session.validate');
         }
+
+        // Set default route name for navigation helpers.
+        Navigation\Page\Mvc::setDefaultRoute('default');
+
+        // Determine the user's role.
+        $groupResultSet = $db->query('SELECT id, groupname FROM scagroup', []);
+        $groupList = [];
+        foreach ($groupResultSet as $group) {
+            $groupList[$group->id] = strtolower(str_replace(' ', '', $group->groupname));
+        }
+
+        $authService = $serviceManager->get(AuthenticationServiceInterface::class);
+        $identity = $authService->getIdentity();
+        $role = 'guest';
+        if ($identity != null) {
+            if ($identity == 'seneschal') {
+                $role = 'admin';
+            } elseif ($identity == 'servers') {
+                $role = 'admin';
+            } elseif (in_array($identity, $groupList)) {
+                $role = 'seneschal';
+            }
+        }
+
+        // Inject ACL into navigation helpers.
+        NavigationHelper::setDefaultAcl($serviceManager->get(AclInterface::class));
+        NavigationHelper::setDefaultRole($role);
     }
 
     public function getConfig()